requirements.txt
pyserial==3.4
setuptools~=68.2.0
websockets~=12.0